e2e_tests/requirements.txt
web3>=6.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
